[#42] Draft of a public endpoint as a landing page for browsers
